refactor: renamed schemas_v2 to schemas
