+++ api/geochronology.py
@@ -17,7 +17,7 @@ from db.geochronology import GeochronologyAge
 from fastapi import APIRouter, Depends, status
 from db import adder
 from db.engine import get_db_session
-from schemas_v2.geochronology import CreateGeochronologyAge
+from schemas.geochronology import CreateGeochronologyAge
 from sqlalchemy.orm import Session
 from sqlalchemy import select
