+++ schemas/asset.py
