+++ schemas/contact.py
@@ -20,8 +20,8 @@ from email_validator import validate_email, EmailNotValidError
 from phonenumbers import NumberParseException
 from pydantic import field_validator, BaseModel, AwareDatetime
 
-from schemas_v2 import ORMBaseModel
-from schemas_v2.thing import ThingResponse
+from schemas import ORMBaseModel
+from schemas.thing import ThingResponse
 
 """
 REFACTOR TODO
