+++ schemas/geothermal.py
