+++ schemas/sensor.py
