+++ schemas/series.py
