+++ schemas/thing.py
@@ -18,8 +18,8 @@ from typing import List
 
 from pydantic import BaseModel, model_validator
 
-from schemas_v2 import ORMBaseModel
-from schemas_v2.location import LocationResponse
+from schemas import ORMBaseModel
+from schemas.location import LocationResponse
 
 
 # -------- CREATE ----------
